+++ CHANGELOG.md
@@ -3,6 +3,16 @@ FlatLaf Change Log
 
 ## 1.3-SNAPSHOT
 
+#### New features and improvements
+
+- TextComponents, ComboBox and Spinner: Support different background color when
+  component is focused (use UI values `TextField.focusedBackground`,
+  `PasswordField.focusedBackground`, `FormattedTextField.focusedBackground`,
+  `TextArea.focusedBackground`, `TextPane.focusedBackground`,
+  `EditorPane.focusedBackground`, `ComboBox.focusedBackground`,
+  `ComboBox.buttonFocusedBackground`, `ComboBox.popupFocusedBackground` and
+  `Spinner.focusedBackground`). (issue #335)
+
 #### Fixed bugs
 
 - Fixed white lines at bottom and right side of window (in dark themes on HiDPI